day 17 + stubs for day18

+++ src/main/java/com/adventofcode/Day15.java
@@ -22,7 +22,7 @@ public class Day15 {
         //int considerRow = 10;
         return getExcludesAndBeacons(considerRow, Integer.MIN_VALUE, Integer.MAX_VALUE).excludes()
                 .stream()
-                .mapToInt(Range::size)
+                .mapToLong(Range::size)
                 .sum();
     }
 
@@ -32,9 +32,9 @@ public class Day15 {
         //final int minX = 0, minY = 0, maxX = 20, maxY = 20;
         for (int considerRow = minY; considerRow <= maxY; considerRow++) {
             ExcludesAndBeacons excludesAndBeacons = getExcludesAndBeacons(considerRow, minX, maxX);
-            int excludesLength = excludesAndBeacons.excludes().stream().mapToInt(Range::size).sum();
+            long excludesLength = excludesAndBeacons.excludes().stream().mapToLong(Range::size).sum();
             if (excludesLength + excludesAndBeacons.beacons().size() < maxX - minX + 1) {
-                for (int i = minX; i <= maxX; i++) {
+                for (long i = minX; i <= maxX; i++) {
                     if (Range.rangesDontContain(excludesAndBeacons.excludes(), i) && !excludesAndBeacons.beacons().contains(i)) {
                         return 4000000L * i + considerRow;
                     }
@@ -45,22 +45,22 @@ public class Day15 {
     }
 
     private ExcludesAndBeacons getExcludesAndBeacons(int considerRow, int minX, int maxX) {
-        Set<Integer> beaconsAtGivenRow = new HashSet<>();
+        Set<Long> beaconsAtGivenRow = new HashSet<>();
         Set<Range> excludes = new HashSet<>();
         for (XY[] xies : input) {
             if (xies[0].y() == considerRow || xies[1].y() == considerRow) {
                 beaconsAtGivenRow.add(xies[1].x());
             }
-            int noBeaconsInDistance = xies[0].manhattanDistance(xies[1]);
-            int[] excludeAllXsInRange = xies[0].findXsOfAllPointsInGivenDistance(noBeaconsInDistance, considerRow);
+            long noBeaconsInDistance = xies[0].manhattanDistance(xies[1]);
+            long[] excludeAllXsInRange = xies[0].findXsOfAllPointsInGivenDistance(noBeaconsInDistance, considerRow);
             if (excludeAllXsInRange.length == 1) {
                 if (excludeAllXsInRange[0] >= minX && excludeAllXsInRange[0] <= maxX) {
                     excludes = new Range(excludeAllXsInRange[0], excludeAllXsInRange[0]).addToSetOfRangesAndCompact(excludes);
                 }
             }
             if (excludeAllXsInRange.length == 2) {
-                int from = Math.max(excludeAllXsInRange[0], minX);
-                int to = Math.min(excludeAllXsInRange[1], maxX);
+                long from = Math.max(excludeAllXsInRange[0], minX);
+                long to = Math.min(excludeAllXsInRange[1], maxX);
                 if (to >= from)
                     excludes = new Range(from, to).addToSetOfRangesAndCompact(excludes);
             }
@@ -69,13 +69,13 @@ public class Day15 {
         return new ExcludesAndBeacons(excludesWithoutBeacons, beaconsAtGivenRow);
     }
 
-    private Set<Range> removeBeacons(Set<Range> excludes, Set<Integer> beaconsAtGivenRow) {
-        LinkedList<Integer> copyOfBeacons = beaconsAtGivenRow.stream().sorted().collect(Collectors.toCollection(LinkedList::new));
+    private Set<Range> removeBeacons(Set<Range> excludes, Set<Long> beaconsAtGivenRow) {
+        LinkedList<Long> copyOfBeacons = beaconsAtGivenRow.stream().sorted().collect(Collectors.toCollection(LinkedList::new));
         return excludes.stream().sorted()
                 .flatMap(range -> {
                     List<Range> ranges = new ArrayList<>();
                     while (range != null && !copyOfBeacons.isEmpty() && range.containsValue(copyOfBeacons.getFirst())) {
-                        Integer beaconPosition = copyOfBeacons.removeFirst();
+                        Long beaconPosition = copyOfBeacons.removeFirst();
                         if (beaconPosition == range.fromInclusive()) {
                             Range r = new Range(range.fromInclusive() + 1, range.toInclusive());
                             range = r.isCorrect() ? r : null;
@@ -98,6 +98,6 @@ public class Day15 {
 
     }
 
-    private record ExcludesAndBeacons(Set<Range> excludes, Set<Integer> beacons) {
+    private record ExcludesAndBeacons(Set<Range> excludes, Set<Long> beacons) {
     }
 }
